+++ src/common/index.ts
@@ -3,4 +3,5 @@
  */
 export const pureObject = <T>(obj: T): T => JSON.parse(JSON.stringify(obj)) as T
 
+export * from './entitlement'
 export * from './sqlParser'